+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - **NEW MINT-2176** Make SDK available on composer - *Joakim.L*
 - **FIX MINT-2089** Fix URI for BETA mode with checkoutService - *Joakim.L*
 - **NEW MINT-2193** Remove checkoutHTML method - *Joakim.L*
+- **NEW MINT-2133** Add function extendExpiryDate - *Joakim.L*
 
 ## v3.2.0 - 2014-11-11
 - **NEW MINT-1894** Add setter for client IP - *Christer.G*